+++ src/sound/SoundDevice.cc
@@ -182,7 +182,7 @@ void SoundDevice::muteChannel(unsigned channel, bool muted)
 bool SoundDevice::mixChannels(int* dataOut, unsigned samples)
 {
 #ifdef __SSE2__
-	assert((long(dataOut) & 15) == 0); // must be 16-byte aligned
+	assert((uintptr_t(dataOut) & 15) == 0); // must be 16-byte aligned
 #endif
 	if (samples == 0) return true;
 	unsigned outputStereo = isStereo() ? 2 : 1;
